rename feature 'transcription' to 'subtitles'
